did lots of stuff today. Changed how we look at ST operations, and added a working stack finally :) The stack operations will probably be moved to their own method eventually... not sure how much that matters yet. The previous ST operations were causing some problems so I had to move them outside of our mathops. The only case handled with this stack is the one from the simpleExpression.tm file :   y = 2*x - 2*y;
The stack code will need to be applied to the other operations to remain consistent. I believe this should be easy since we have our crazy IF STATEMENTS that check for every possibility, so copy/pasting into the appropriate section should suffice. A division operator is still necessary as well.

Priority needs to be working on a functional lineCount offset. If we solve this roadblock the rest of this stage shouldn't be difficult since we now have a functional stack.

Final note: With how we use our stack, I am certain the way we assign addresses to variables will need to be  changed. This causes no problems now but may in larger programs. If you see any code you have questions about let me know, since I may have forgotten to comment on them in this message.